Added bot messages and auto-leave when song finishes
When a song finishes, the callback now dispatches a coroutine callback
in the event loop of the main thread. In the callback, the bot posts a
message to the channel it was invited through with the title of the next
song. If no more songs are scheduled, it automatically leaves the
channel (but only if it did't explicitly join through !join).