Riffbot now sets the current song as channel topic
If the manage_channels permission is given, the bot will now set the
channel topic of the text channel through which it is receiving commands
to the song that is currently playing. It will no longer send a message
to the channel every time a new song starts. Play/pause status is now
also reflected in the channel topic.

Furthermore, the bot now shows a typing indicator while it is searching
YouTube for the requested songs on !play.