Added config option support
Configuration options can now be specified in an ini-formatted file.
This file can be passed to riffbot by starting it with the
"-c <filepath>" command line option. All options that work on the
command line can also be specified in the config file under the
[default] section (omitting the leading "--"). If an option is present
both on the command line and in the config file, the comand line value
takes precedence.

Also added a new option "--command-prefix" that allows you to specify a
custom command prefix (default: !).